requirements.txt
aiohttp==3.8.1
paho-mqtt
python-dotenv